App plots (some times) the required data.
The plotting depends on the value of the variable loaded. Its
initialized as false and when the data are ready it get true value,
while the component of the chart Bar is checking if the variable turned
to value of true. The problem here is that after the first plotting of
data, the loaded has a value of true in every case. Thus the component
is rendered at the start of the session, where the data are not yet fetched